Added tmux-MacOSX-pasteboard by @ChrisJohnsen
This enables pbcopy and * register usage for Vim within tmux on
Mac OS X. To be able use tmux even if this wrapper script is
not installed the clipboard_for_tmux_osx wraps the call to it
and just calls bash if reattach-to-user-namespace does not exist.

The pasteboard wrapper can be installed with `rake
tmux:build_pasteboard_wrapper` and deinstalled with `rake
tmux:clean_pasteboard_wrapper`.