Got rid of Ruby/Rake dependency for installing the dotfiles
The installation is now done via a bash script that has fewer
dependencies. Also, the linking strategy is now a bit more flexible:
Only files on the first hierachy of this repository that have a filename
starting with an underscore will be linked as a dotfile to the home
directory.